% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_functions.R
\name{add_cov_stats_ind}
\alias{add_cov_stats_ind}
\title{Adds columns specifying true mu, 95\% coverage, bias and precision to the cmdstanr fit summary dataframe containing only individual parameter rows}
\usage{
add_cov_stats_ind(ind_fit, ind_pars, model_num)
}
\arguments{
\item{ind_fit}{filtered cmdstanr summary fit dataframe (only individual par rows)}

\item{ind_pars}{list containing true individual parameters}

\item{model_num}{(int) specifying the number of stan models used in the dataframe}
}
\description{
Adds columns specifying true mu, 95\% coverage, bias and precision to the cmdstanr fit summary dataframe containing only individual parameter rows
}
