% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_functions.R
\name{add_cov_stats_mu}
\alias{add_cov_stats_mu}
\title{Adds columns specifying true mu, 95\% coverage, bias and precision to the cmdstanr fit summary dataframe containing only mu rows}
\usage{
add_cov_stats_mu(mu_fit, mu_list, model_num = 1)
}
\arguments{
\item{mu_fit}{filtered cmdstanr summary fit dataframe (only mu rows)}

\item{mu_list}{named list containing true mu values}

\item{model_num}{(int) specifying the number of stan models used in the dataframe}
}
\description{
Adds columns specifying true mu, 95\% coverage, bias and precision to the cmdstanr fit summary dataframe containing only mu rows
}
