% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_prepro.R
\name{add_fixed_pars}
\alias{add_fixed_pars}
\title{add fixed pars to stan list}
\usage{
add_fixed_pars(stan_list, fixed_pars, model_struct_list)
}
\arguments{
\item{model_struct_list}{}
}
\description{
add fixed pars to stan list
}
