% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_prepro.R
\name{add_gammas}
\alias{add_gammas}
\title{Add gamma pars from model_struct_list to stan_list}
\usage{
add_gammas(stan_list, model_struct_list)
}
\arguments{
\item{model_struct_list}{}
}
\description{
Add gamma pars from model_struct_list to stan_list
}
