% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_summaries.R
\name{add_vars_from_stan_list_to_model_output}
\alias{add_vars_from_stan_list_to_model_output}
\title{Adds variables from stan_list to to stan model summary (3-D, 2-D or 1-D)}
\usage{
add_vars_from_stan_list_to_model_output(
  exp_struct_list,
  cmdstan_summary,
  stan_list,
  var_names
)
}
\arguments{
\item{exp_struct_list}{has to have 'ntrial' and 'nblock' field, specifying the number of trials per subject}

\item{cmdstan_summary}{summary object}

\item{stan_list}{stan list containing the variables to add}

\item{var_names}{names of variables to add}
}
\description{
Useful for supplying simulated model with simulated parameters and experimental vars from task skeleton
}
