% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_distribution.R
\name{beta_mean}
\alias{beta_mean}
\title{Estimate mean of the beta distribution}
\usage{
beta_mean(alpha, beta)
}
\arguments{
\item{beta}{}
}
\value{
Mean of the Beta distribution
}
\description{
Estimate mean of the beta distribution
}
