% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_distribution.R
\name{beta_sd}
\alias{beta_sd}
\title{Estimate the standard deviation of the Beta distribution}
\usage{
beta_sd(alpha, beta)
}
\arguments{
\item{alpha}{Alpha parameter of the Beta distribution}

\item{beta}{Beta parameter of the Beta distribution}
}
\description{
Estimate the standard deviation of the Beta distribution
}
