% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_functions.R
\name{calc_loo}
\alias{calc_loo}
\title{Calculate LOOIC of a cmdstan model object}
\usage{
calc_loo(fit)
}
\arguments{
\item{fit}{either draws or a character vector which can be retrieved using targets}
}
\description{
Calculate LOOIC of a cmdstan model object
}
