% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_model_comp.R
\name{calc_waic}
\alias{calc_waic}
\title{Calculate WAIC (and llpd) of a cmdstan model object}
\usage{
calc_waic(fit)
}
\arguments{
\item{fit}{either draws or a character vector which can be retrieved using targets}
}
\description{
Calculate WAIC (and llpd) of a cmdstan model object
}
