% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_prepro.R
\name{compile_model}
\alias{compile_model}
\title{Compiles a stan model (wrapper for cmdstanr::cmdstanmodel())}
\usage{
compile_model(model_file)
}
\arguments{
\item{model_file}{}
}
\description{
Compiles a stan model (wrapper for cmdstanr::cmdstanmodel())
}
