% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decay.R
\name{decay_exp}
\alias{decay_exp}
\title{Exponential Decay}
\usage{
decay_exp(A0, k, t)
}
\arguments{
\item{A0}{Initial value}

\item{k}{Decay slope (0: immediate decay; 1:no decay)}

\item{t}{Time step}
}
\description{
Exponential Decay
}
