% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decay.R
\name{decay_hyper}
\alias{decay_hyper}
\title{Hyperbolic decay}
\usage{
decay_hyper(A0, k, t)
}
\arguments{
\item{A0}{Initial value}

\item{k}{Decay slope (0: no decay)}

\item{t}{Time step}
}
\description{
Hyperbolic decay
}
