% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decay.R
\name{decay_qhyper}
\alias{decay_qhyper}
\title{Quasi-hyperbolic decay}
\usage{
decay_qhyper(A0, k_qexp, k_qhyper, t)
}
\arguments{
\item{A0}{Initial value}

\item{k_qexp}{Exponential decay component}

\item{k_qhyper}{Hyperbolic Decay component}

\item{t}{Time step}
}
\description{
Quasi-hyperbolic decay
}
