% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_parameters2.R
\name{draw_mu_from_gamma}
\alias{draw_mu_from_gamma}
\title{Title}
\usage{
draw_mu_from_gamma(
  gamma_parsA = c(1, 1),
  gamma_parsB = c(1, 1),
  shift = 0,
  scale = 1,
  nsim
)
}
\arguments{
\item{gamma_parsA}{gamma priors for alpha parameter of the Beta distribution}

\item{gamma_parsB}{gamma priors for beta parameter of the Beta distribution}

\item{nsim}{}
}
\description{
Title
}
