% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_distribution.R
\name{estBetaParams}
\alias{estBetaParams}
\title{Estimate alpha and beta parameters of the Beta distribution based on the mean and std or variance}
\usage{
estBetaParams(mu, var = NULL, std = NULL)
}
\arguments{
\item{mu}{mean}

\item{var}{variance}

\item{std}{standard deviation}
}
\value{
list of parameters
}
\description{
Estimate alpha and beta parameters of the Beta distribution based on the mean and std or variance
}
