% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_model_comp.R
\name{extract_loos}
\alias{extract_loos}
\title{Extract LOOIC values from all models in your targets folder ('draws' objects)}
\usage{
extract_loos(model_dir = "_targets/objects/", model_str = NULL)
}
\arguments{
\item{model_dir}{Specifies the targets directory (default should work)}

\item{model_str}{Specifies model string}
}
\value{
a list of loo scores
}
\description{
Extract LOOIC values from all models in your targets folder ('draws' objects)
}
