% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_summaries2.R
\name{extract_preds}
\alias{extract_preds}
\title{Extracts simulated pars}
\usage{
extract_preds(cmdstan_summary, preds = c("d_pred", "r_pred"))
}
\arguments{
\item{preds}{a list of predictor names}
}
\description{
Extracts simulated pars
}
