% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_fit.R
\name{fit_model}
\alias{fit_model}
\title{Fit a stan model with cmdstanr default fitting setup}
\usage{
fit_model(
  stan_data,
  model_file,
  summary_only = T,
  iter_warmup = 2000,
  iter_sampling = 2000,
  chains = 2,
  parallel_chains = 2,
  refresh = 500,
  seed = 1,
  ...
)
}
\arguments{
\item{stan_data}{A \code{list} with data, formatted for the stan model}

\item{model_file}{(char) model path}

\item{summary_only}{(bool)}

\item{iter_warmup}{(int) warmup interations}

\item{iter_sampling}{(int) sampling iterations}

\item{chains}{(int) number of chains}

\item{parallel_chains}{(int) how many chains in parallel}

\item{refresh}{(int) how often to print info}

\item{seed}{(int) seed}
}
\description{
A wrapper around \code{cmdstanr::cmdstan_model}
}
