% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_fit.R
\name{fit_model_gen}
\alias{fit_model_gen}
\title{Simulate a synthetic dataset using a stan model and a parameter list}
\usage{
fit_model_gen(stan_data, model_file, summary_only = T)
}
\arguments{
\item{stan_data}{stan list containing task skeleton + individual parameter values}

\item{model_file}{.stan file with generated quantities block}

\item{summary_only}{}
}
\description{
Simulate a synthetic dataset using a stan model and a parameter list
}
