% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_summaries.R
\name{get_inds}
\alias{get_inds}
\title{Filters cmdstan fit summary dataframe for mu values and adds fit diagnostic columns (add_cov_stats_ind())}
\usage{
get_inds(
  stan_fit_summary,
  ind_pars,
  model_num,
  param_list = NULL,
  many_models = F
)
}
\arguments{
\item{stan_fit_summary}{cmdstan fit summary dataframe}

\item{ind_pars}{list containing true individual parameters}

\item{model_num}{(int) specifying the number of stan models used in the dataframe}
}
\description{
Filters cmdstan fit summary dataframe for mu values and adds fit diagnostic columns (add_cov_stats_ind())
}
