% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_summaries2.R
\name{get_inds2}
\alias{get_inds2}
\title{Extract individual parameters from cmdstanr model summary object}
\usage{
get_inds2(stan_fit_summary, ind_pars = NULL)
}
\arguments{
\item{ind_pars}{}
}
\description{
Extract individual parameters from cmdstanr model summary object
}
