% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_summaries.R
\name{get_mus}
\alias{get_mus}
\title{Filters cmdstan fit summary dataframe for mu values and adds fit diagnostic columns (add_cov_stats_mu())}
\usage{
get_mus(stan_fit_summary, mu_list, model_num = 1)
}
\arguments{
\item{stan_fit_summary}{cmdstan fit summary dataframe}

\item{mu_list}{named list containing true mu values}

\item{model_num}{(int) specifying the number of stan models used in the dataframe}
}
\description{
Filters cmdstan fit summary dataframe for mu values and adds fit diagnostic columns (add_cov_stats_mu())
}
