% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_summaries2.R
\name{get_mus2}
\alias{get_mus2}
\title{Extract group-level mean parameters (mus) from cmdstanr model summary object}
\usage{
get_mus2(stan_fit_summary)
}
\arguments{
\item{stan_fit_summary}{}
}
\description{
Extract group-level mean parameters (mus) from cmdstanr model summary object
}
