% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get_sim_mu_sds}
\alias{get_sim_mu_sds}
\title{Calculate sd's for each group parameter}
\usage{
get_sim_mu_sds(mu_list, sim_data)
}
\arguments{
\item{mu_list}{named list of group-level  parameters (names should start with 'mu_')}

\item{sim_data}{simulated datasets with varying levels of parameter sets}
}
\description{
Calculate sd's for each group parameter
}
