% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decay.R
\name{growth_hyper}
\alias{growth_hyper}
\title{Hyperbolic growth}
\usage{
growth_hyper(A0, k, t)
}
\arguments{
\item{A0}{Initial value}

\item{k}{Exponential growth (0: no growth)}

\item{t}{Time step}
}
\description{
Hyperbolic growth
}
