% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning.R
\name{kalman}
\alias{kalman}
\title{Kalman filter}
\usage{
kalman(vals, vars, num_chosen, reward, sigma_noise, sigma_change)
}
\arguments{
\item{vals}{vector of 2 values}

\item{vars}{vector od 2 standard deviations}

\item{num_chosen}{int specifying which option was chosen (1 or 2)}

\item{reward}{reward value}

\item{sigma_noise}{estimation noise parameter}

\item{sigma_change}{environmental stability parameter}
}
\description{
Kalman filter
}
