% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{loadRData}
\alias{loadRData}
\title{Rdata Loader}
\usage{
loadRData(fileName)
}
\arguments{
\item{fileName}{Rdata object}
}
\description{
Loads loads an RData file and returns it (can assign it to a new variable)
}
