% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_prepro.R
\name{make_fixed_pars_list}
\alias{make_fixed_pars_list}
\title{Creates a list of fixed parameters}
\usage{
make_fixed_pars_list(parameter_list, fix_list, non_fixed_number_code = -1)
}
\arguments{
\item{parameter_list}{a list containing all parameter names (often range_list or mu_list can be used)}

\item{fix_list}{a list contatining only fixed parameter names and values, e.g. parX = 4}

\item{non_fixed_number_code}{-1 by default (needs to be numeric for STAN)}
}
\description{
Outputs a list of all the model parameters with 'fix_' prefix, specifying which parameters are Fixed. By default, -1 indicates non-fixed, while all other values specify the fixed value. Parameter names taken from 'paramater_list' (mu_list or range_list can be used). Fixed vals taken from 'fixed list'.
}
