% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_build.R
\name{make_stan_skeleton}
\alias{make_stan_skeleton}
\title{Builds a stan-model skeleton, specifying priors, their ranges, and shape, based on the idea of sampling from a bounded continuous space}
\usage{
make_stan_skeleton(var_names)
}
\arguments{
\item{var_names}{list of parameter names}
}
\description{
Builds a stan-model skeleton, specifying priors, their ranges, and shape, based on the idea of sampling from a bounded continuous space
}
