% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_model_comp.R
\name{model_comp_table}
\alias{model_comp_table}
\title{Creates a table of loo/waic scores based on loo lists from \code{extract_loos} and \code{extract_waics} functions}
\usage{
model_comp_table(
  loo_list = NULL,
  waic_list = NULL,
  model_str = NULL,
  relevant_pars = "looic"
)
}
\arguments{
\item{loo_list}{a list of model loo values created with \code{extract_loos} function}

\item{waic_list}{a list of model WAIC values created with \code{extract_waics} function}

\item{model_str}{Specifies model string}

\item{relevant_pars}{character string (seems superfluous, since the function should automatically detect which parameters to extract based on the first 2 arguments)}
}
\description{
Creates a table of loo/waic scores based on loo lists from \code{extract_loos} and \code{extract_waics} functions
}
