% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_mu_on_acc}
\alias{plot_mu_on_acc}
\title{Interatively plots task accuracies based on different levels of simulated mu parameters}
\usage{
plot_mu_on_acc(mu_list, sim_data)
}
\arguments{
\item{mu_list}{named list of group-level  parameters (names should start with 'mu_')}

\item{sim_data}{simulated datasets with varying levels of parameter sets}
}
\description{
Interatively plots task accuracies based on different levels of simulated mu parameters
}
