% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{setGGoptions}
\alias{setGGoptions}
\title{Set GG options}
\usage{
setGGoptions(base = 10, line_wid = 1)
}
\arguments{
\item{base}{letter size}

\item{line_wid}{line width}
}
\description{
Set GG options
}
