% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{setOptions}
\alias{setOptions}
\title{Setting various options}
\usage{
setOptions(...)
}
\arguments{
\item{...}{parameters passed to specific setXOptions funcs. Currently only for GGoptions.}
}
\value{
NA
}
\description{
Setting various options
}
