% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{sigm_drift}
\alias{sigm_drift}
\title{Sigmoidal drift-rate transform (based on: https://pubmed.ncbi.nlm.nih.gov/30924057/)}
\usage{
sigm_drift(z, vmax)
}
\arguments{
\item{z}{original drift rate}

\item{vmax}{maximum drift rate (abs)}
}
\description{
Sigmoidal drift-rate transform (based on: https://pubmed.ncbi.nlm.nih.gov/30924057/)
}
