% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_parameters.R
\name{sim_ind_pars}
\alias{sim_ind_pars}
\title{Simulate individual parameters from beta-like group-level priors (with lower and upper bounds)}
\usage{
sim_ind_pars(
  mu_list,
  sigma_list = NULL,
  range_list = NULL,
  sigma = 0.1,
  range = c(0, 1),
  nsub,
  seed = NULL
)
}
\arguments{
\item{mu_list}{A \code{list} of group-level means. Each element needs to be named as the individual parameter.}

\item{sigma_list}{(optional) A \code{list} of group-level standard deviation parameters, in beta distribution-space. If all sigmas are the same, this argument can stay as \code{NULL}}

\item{range_list}{(optional) A \code{list} of group-level lower and upper bounds. If all bounds are the same, this argument can be skipped}

\item{sigma}{(optional) Single value specifying all group-level standard deviation parameters}

\item{range}{(optional) Single \code{list} specifying all group-level parameter ranges (\code{c(lb, ub)} format).}

\item{nsub}{number of participants}

\item{seed}{self-explanatory}
}
\value{
A \code{list} containing individual parameters
}
\description{
The function takes in group-level means ('mu' parameters), standard deviations ('sigmas')
}
