% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_parameters.R
\name{sim_ind_pars_multi}
\alias{sim_ind_pars_multi}
\title{Simulate individual parameters from beta-like group-level priors (with lower and upper bounds) for multiple simulations with different parameter settings}
\usage{
sim_ind_pars_multi(
  exp_struct_list,
  mu_list,
  sigma_list = NULL,
  range_list = NULL,
  sigma = 0.1,
  range = c(0, 1),
  seed = NULL
)
}
\arguments{
\item{exp_struct_list}{list specifying the experimental structure (current version needs only the nsub field, specifying the number of participants)}

\item{mu_list}{hierarchical list specifying group parameter means (e.g.\code{list(mu_alpha=c(0.25,0.5,0.75),mu_tau  =c(0.25,0.5,0.75))}}

\item{seed}{you know what a seed is}
}
\description{
Simulate individual parameters from beta-like group-level priors (with lower and upper bounds) for multiple simulations with different parameter settings
}
