% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_parameters.R
\name{sim_mus}
\alias{sim_mus}
\title{Simulate group means (mus)  from beta-like priors}
\usage{
sim_mus(
  mu_list,
  sigma_list = NULL,
  range_list = NULL,
  sigma = 0.1,
  range = c(0, 1),
  nsim,
  seed = NULL
)
}
\arguments{
\item{seed}{}
}
\description{
Simulate group means (mus)  from beta-like priors
}
