% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_parameters2.R
\name{sim_mus2}
\alias{sim_mus2}
\title{Title}
\usage{
sim_mus2(model_struct_list, prior_dist = "gamma", nsim, seed = NULL)
}
\arguments{
\item{model_struct_list}{Model struture list}

\item{prior_dist}{prior distribution (currently works for 'gamma' (default), 'beta' and 'gauss')}

\item{nsim}{number of simulations}

\item{seed}{}
}
\description{
Title
}
