% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_parameters2.R
\name{sim_raw_inds_from_ab}
\alias{sim_raw_inds_from_ab}
\title{simulates raw individual parameter values (in beta space) from alpha/beta dataframes}
\usage{
sim_raw_inds_from_ab(
  mus_df,
  out_pars = "beta",
  nsim,
  nsubj,
  starting_string = "",
  seed = NULL
)
}
\arguments{
\item{starting_string}{string at the start of each column name. Defaults to '' (empty), but sometimes 'mu_' might work}

\item{seed}{}
}
\description{
simulates raw individual parameter values (in beta space) from alpha/beta dataframes
}
