% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_parameters2.R
\name{sim_raw_pars_from_ab}
\alias{sim_raw_pars_from_ab}
\title{Simulates raw group parameter values (in beta space) from alpha/beta dataframes}
\usage{
sim_raw_pars_from_ab(
  settings_list,
  alpha_df,
  beta_df,
  out_pars = "beta",
  single_df = T,
  seed = NULL
)
}
\arguments{
\item{settings_list}{has to contain $nsim field with number of simulations}

\item{alpha_df}{N-by-p df, where p are the parameter names (in the form 'a_x_raw')}

\item{beta_df}{same for beta values ('b_x_raw' colnames)}

\item{out_pars}{output parametrization: 'beta' returns alpha & beta values; otherwise: mean + sd}

\item{single_df}{output type: TRUE returns a single dataframe; FALSE a list of 2 elements}

\item{seed}{}
}
\description{
Parameter recovery: for drawing group-level parameter values
}
