% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_subj_sim.R
\docType{data}
\name{single_subj_sim}
\alias{single_subj_sim}
\title{Single subject data simulation}
\format{
A \code{tibble} with 4 columns, which are:
\describe{
\item{cond}{Difficulty condition ('Easy' or 'Difficult')}
\item{trial}{trial number per difficulty condition}
\item{choice}{Accuracy (binary)}
\item{rt}{Reaction Time (s), bound between 0.2 and 3 (simulated from a truncated normal distribution)}
}
}
\usage{
single_subj_sim
}
\description{
Simulated data \code{tibble} containing a single subject responses (accuracies and rts) in two difficulty conditions
}
\keyword{datasets}
