% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choices.R
\name{softmax}
\alias{softmax}
\title{Softmax function (2 values)}
\usage{
softmax(vals, temperature)
}
\arguments{
\item{vals}{vector of values (length==2)}

\item{temperature}{inverse temperature parameter (higher values -> more precision)}
}
\description{
Softmax function (2 values)
}
