% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choices.R
\name{softmax_inverse}
\alias{softmax_inverse}
\title{Softmax (2 values, with inverted tau)}
\usage{
softmax_inverse(vals, tau)
}
\arguments{
\item{vals}{vector of values (length==2)}

\item{tau}{temperature parameter (higher values -> less precision)}
}
\description{
Softmax (2 values, with inverted tau)
}
