% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{sourceAll}
\alias{sourceAll}
\title{Sources all the code within given folder + file name pattern}
\usage{
sourceAll(path, pattern = "_func")
}
\arguments{
\item{path}{character vector specifying the path to the folder with to-be-sourced code}

\item{pattern}{character vector specifying the file name pattern}
}
\value{
0
}
\description{
Sources all the code within given folder + file name pattern
}
