% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choices.R
\name{thompson}
\alias{thompson}
\title{Thompson sampling (normal or beta (moment-matching method))}
\usage{
thompson(par1, par2, pay_type)
}
\arguments{
\item{par1}{if pay_type == gauss: par1 <- vector of 2 means; if pay_type == beta: par1 <- vector of 2 alphas}

\item{par2}{if pay_type == gauss: par1 <- vector of 2 sds;  if pay_type == beta: par2 <- vector of 2 betas}

\item{pay_type}{payoff type: drawn from either normal ('gauss') or beta ('beta') distribution}
}
\description{
Thompson sampling (normal or beta (moment-matching method))
}
