% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_summaries2.R
\name{widen_preds}
\alias{widen_preds}
\title{Turns multiple variables from a cmdstan_summary df to separate columns using pivot_wider}
\usage{
widen_preds(cmdstan_summary)
}
\arguments{
\item{cmdstan_summary}{}
}
\description{
Turns multiple variables from a cmdstan_summary df to separate columns using pivot_wider
}
