% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_functions.R
\name{make_stan_list}
\alias{make_stan_list}
\title{Making a stan-compatible data list}
\usage{
make_stan_list(
  dat,
  var_names,
  grouping_vars = c("subjID", "game", "trial"),
  grouping_vars_stan_names = c("numSubj", "numGames", "numTrials"),
  hierarchical = T,
  ssm_prepro = F
)
}
\arguments{
\item{dat}{original dataframe}

\item{var_names}{output variables (choice, rt, outcome, etc...)}

\item{grouping_vars}{grouping variables (subjID, games, trials)}

\item{grouping_vars_stan_names}{specifying the grouping variable stan-names}

\item{hierarchical}{whether the dataset is multi-subject (T) or not (F)}

\item{ssm_prepro}{whether to include lower RT bounds for sequential sampling models input}
}
\description{
Making a stan-compatible data list
}
