% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_summaries2.R
\name{pars_to_stan}
\alias{pars_to_stan}
\title{Extracts individual parameter values from the task skeleton back to a list}
\usage{
pars_to_stan(dataframe, mu_list, grouping_cols = c("subjID"))
}
\arguments{
\item{dataframe}{task skeleton}

\item{grouping_cols}{for individual parameters only subjID}
}
\description{
Extracts individual parameter values from the task skeleton back to a list
}
