% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_predictions.R, R/stan_summaries2.R
\name{widen_preds}
\alias{widen_preds}
\title{Widens prediction dataframe (useful when 2+ prediction variables are generated)}
\usage{
widen_preds(cmdstan_summary)

widen_preds(cmdstan_summary)
}
\arguments{
\item{cmdstan_summary}{}

\item{cmdstan_preds}{a cmdstan summary with extracted prediction variables}
}
\description{
Widens prediction dataframe (useful when 2+ prediction variables are generated)

Turns multiple variables from a cmdstan_summary df to separate columns using pivot_wider
}
